+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+ - Command line argument parsing, you can now open a module on application startup
+
 ## [0.5.0] - 2021-09-10
 ### Added
  - Instrument + Waveform list views for selecting/adding instruments/waveforms